plugins: Add AsyncDBusEvalMixin
Regular AsyncEvalMixin can still cause blockage of the main thread
in case it performs a blokcing DBus call and the main thread attempts
to perform it too.

Add a version of AsyncEvalMixin that solves the problem by introducing
callbacks that are to be leveraged by plugin when it performs the
DBus call.